I can't seem to connect them properly T^T
When I tried to connect the share post feature, the dropdown can't be seen even though it's there. I think something is blocking it. As for the upload image feature, I'm can't find the file that contains the "New post" button. The upload image button should be next to the share button.